+++ market/pricingModels/verticalDemandAndStepwiseSupply.js
